Next step of the fire system change
So what's different now is that you need to take the stuff in your hand to ignite it before you store it. Also you have to store the whole pile or otherwise the part that's in your hand remains on fire. If you want to spare material, drop some, then store the part that's on fire.